# api.yaml
openapi: 3.0.3
info:
  title: WireGuard Obfuscator Easy API
  description: |
    REST API for managing WireGuard VPN server with obfuscation support.
    Provides endpoints for authentication, client management, server configuration,
    statistics, and monitoring.
  version: 1.0.0
  contact:
    name: Alexey Cluster
    email: cluster@cluster.wtf
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: /{webPrefix}
    description: Production server with web prefix
    variables:
      webPrefix:
        default: ""
        description: Optional web prefix (e.g., /vpn)
...
paths:
...
  /api/status:
    get:
      tags:
        - Statistics
      summary: Get server status
      description: Get server status including WireGuard and obfuscator status
      operationId: getStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Server status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/stats:
    get:
      tags:
        - Statistics
      summary: Get WireGuard statistics
      description: Get WireGuard statistics for all peers
      operationId: getStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: WireGuard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WireGuardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          description: WireGuard interface not found or not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/stats/{username}:
    get:
      tags:
        - Statistics
      summary: Get client statistics
      description: Get WireGuard statistics for a specific client
      operationId: getClientStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Client statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          description: WireGuard interface not found or not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/logs/obfuscator:
    get:
      tags:
        - Statistics
      summary: Get obfuscator logs
      description: Get last N lines from in-memory obfuscator logs
      operationId: getObfuscatorLogs
      security:
        - bearerAuth: []
      parameters:
        - name: lines
          in: query
          description: Number of lines to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
      responses:
        '200':
          description: Obfuscator logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  lines:
                    type: array
                    items:
                      type: string
                  total_lines:
                    type: integer
                  requested_lines:
                    type: integer
                  returned_lines:
                    type: integer
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication. For most endpoints, use JWT token obtained from the /api/auth/login endpoint.
        For metrics endpoints (/api/metrics/*), you can use either a JWT token or the dedicated metrics token (obtained from /api/metrics/token).
        Include the token in the Authorization header as: "Bearer <token>"
  parameters:
    Username:
      name: username
      in: path
      required: true
      description: Client username
      schema:
        type: string
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    ServerStatus:
      type: object
      properties:
        external_ip:
          type: string
          format: ipv4
        external_port:
          type: integer
        subnet:
          type: string
        server_ip:
          type: string
        server_public_key:
          type: string
        clients_count:
          type: integer
          description: Total number of clients
        connected_clients_count:
          type: integer
          description: Number of currently connected clients
        version:
          type: string
          description: Application version
        wireguard:
          type: object
          properties:
            running:
              type: boolean
            error:
              type: string
              nullable: true
        obfuscator:
          type: object
          properties:
            enabled:
              type: boolean
            running:
              type: boolean
            error:
              type: string
              nullable: true
            exit_code:
              type: integer
              nullable: true
            version:
              type: string
              nullable: true
              description: Obfuscator version
    WireGuardStats:
      type: object
      properties:
        interface:
          type: string
          description: WireGuard interface name
        peers:
          type: array
          items:
            $ref: '#/components/schemas/PeerStats'
    PeerStats:
      type: object
      properties:
        public_key:
          type: string
        client_name:
          type: string
          nullable: true
        is_connected:
          type: boolean
        latest_handshake:
          type: integer
          nullable: true
          description: Unix timestamp of last handshake
        rx_bytes:
          type: integer
          description: Total received bytes
        tx_bytes:
          type: integer
          description: Total sent bytes
        message:
          type: string
          nullable: true
          description: Status message (e.g., "Client not currently connected")
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'